Add files via upload
Java Products Database Application was created for an assignment while I was studying Software Development. It was made under a scenario for a fictional company called "One World Health". The system allows the user to add, edit and delete products from the system, display the products by category and also make sales as there is an add to basket and checkout feature.